Use PrettyText components
